requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
supabase-py
celery[redis]
pandas
openpyxl
docxtpl
python-multipart
pymysql 
sqlalchemy 
cryptography